Add privacy accounting to dp scaffold example

+++ fl4health/privacy/fl_accountants.py
@@ -57,7 +57,7 @@ class FlInstanceLevelAccountant:
         """server_updates: number of central server updates performed"""
         epsilons = []
         for num_batch, sampling_strategy in zip(self.num_batches_per_client, self.sampling_strategies_per_client):
-            total_updates = server_updates * self.epochs_per_round * num_batch
+            total_updates = int(server_updates * self.epochs_per_round * num_batch)
             epsilon = self.accountant.get_epsilon(sampling_strategy, self.noise_multiplier, total_updates, delta)
             epsilons.append(epsilon)
         return max(epsilons)
@@ -66,7 +66,7 @@ class FlInstanceLevelAccountant:
         """server_updates: number of central server updates performed"""
         deltas = []
         for num_batch, sampling_strategy in zip(self.num_batches_per_client, self.sampling_strategies_per_client):
-            total_updates = server_updates * self.epochs_per_round * num_batch
+            total_updates = int(server_updates * self.epochs_per_round * num_batch)
             delta = self.accountant.get_delta(sampling_strategy, self.noise_multiplier, total_updates, epsilon)
             deltas.append(delta)
         return max(deltas)
